+++ package.json
@@ -52,7 +52,7 @@
     "drizzle-orm": "^0.28.6",
     "drizzle-zod": "^0.5.1",
     "lucide-react": "^0.292.0",
-    "next": "^14.0.1",
+    "next": "14.0.1",
     "next-auth": "^4.24.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
