+++ src/app/(public)/[category]/page.tsx
@@ -1,14 +1,9 @@
-import dynamicImport from "next/dynamic";
 import { notFound } from "next/navigation";
 import { Suspense } from "react";
 import { categoryToTitleMap } from "~/constants";
 import { itemCategory, type ItemCategory } from "~/server/db/schema";
 import { ItemCardRow } from "./_components/item-card-row";
-
-const ClaimModal = dynamicImport(
-  () => import("./_components/claim-modal").then((e) => e.ClaimModal),
-  { ssr: false },
-);
+import { ClaimModalClientWrapper } from "./claim-modal-wrapper";
 
 export default function Page({ params }: { params: Params }) {
   if (!itemCategory.enumValues.includes(params.category as ItemCategory))
@@ -37,9 +32,7 @@ export default function Page({ params }: { params: Params }) {
           itemType="base"
         />
       </Suspense>
-      <Suspense>
-        <ClaimModal />
-      </Suspense>
+      <ClaimModalClientWrapper />
     </main>
   );
 }
